Second stage of restructuring
