+++ ssdaq/bin/terminaldumpers.py
@@ -1,5 +1,4 @@
-from ssdaq import SSReadoutSubscriber
-from ssdaq.subscribers import basicsubscriber
+from ssdaq import subscribers
 from ssdaq.utils import common_args as cargs
 from ssdaq import sslogger
 import logging
@@ -35,7 +34,7 @@ def slowsignaldump():
     args = parser.parse_args()
     eval("sslogger.setLevel(logging.%s)" % args.verbose)
 
-    rsub = SSReadoutSubscriber(port=args.sub_port, ip=args.sub_ip)
+    rsub = subscribers.SSReadoutSubscriber(port=args.sub_port, ip=args.sub_ip)
     rsub.start()
 
     readout_counter = np.zeros(32)
@@ -96,7 +95,7 @@ def timestampdump():
     args = parser.parse_args()
     eval("sslogger.setLevel(logging.%s)" % args.verbose)
 
-    sub = basicsubscriber.BasicTimestampSubscriber(port=args.sub_port, ip=args.sub_ip)
+    sub = subscribers.BasicTimestampSubscriber(port=args.sub_port, ip=args.sub_ip)
     sub.start()
 
     signal.alarm(0)
@@ -149,7 +148,7 @@ def triggerdump():
     args = parser.parse_args()
     eval("sslogger.setLevel(logging.%s)" % args.verbose)
 
-    sub = basicsubscriber.BasicTriggerSubscriber(port=args.sub_port, ip=args.sub_ip)
+    sub = subscribers.BasicTriggerSubscriber(port=args.sub_port, ip=args.sub_ip)
     sub.start()
 
     signal.alarm(0)
@@ -222,7 +221,7 @@ def logdump():
     args = parser.parse_args()
     eval("sslogger.setLevel(logging.%s)" % args.verbose)
 
-    sub = basicsubscriber.BasicLogSubscriber(port=args.sub_port, ip=args.sub_ip)
+    sub = subscribers.BasicLogSubscriber(port=args.sub_port, ip=args.sub_ip)
     sub.start()
 
     signal.alarm(0)
@@ -255,7 +254,7 @@ def mondumper():
     args = parser.parse_args()
     eval("sslogger.setLevel(logging.%s)" % args.verbose)
 
-    sub = basicsubscriber.BasicMonSubscriber(port=args.sub_port, ip=args.sub_ip)
+    sub = subscribers.BasicMonSubscriber(port=args.sub_port, ip=args.sub_ip)
     sub.start()
 
     signal.alarm(0)
