--- ssdaq/core/io/__init__.py
@@ -1,10 +1,0 @@
-from .protobuf_io import RawObjectWriterBase
-from .protobuf_io import RawObjectReaderBase
-from .protobuf_io import LogWriter
-from .protobuf_io import LogReader
-from .protobuf_io import TimestampWriter
-from .protobuf_io import TimestampReader
-from .protobuf_io import TriggerWriter
-from .protobuf_io import TriggerReader
-from .slowsignal_io import SSDataWriter
-from .slowsignal_io import SSDataReader
