+++ ssdaq/data/slowsignal_format.py
