+++ ssdaq/data/trigger_format.py
