+++ ssdaq/receivers/trigger_receiver.py
@@ -1,6 +1,6 @@
 import asyncio
-from ssdaq.core.receiver_server import ReceiverServer, ReceiverMonSender
-
+from ssdaq.core.receiver_server import ReceiverServer
+from .mon_sender import ReceiverMonSender
 
 class TriggerPacketProtocol(asyncio.Protocol):
     def __init__(self, loop, log):
